Rename package from quantfolio to corefolio and update related references in README, code, and tests.
