+++ corefolio/__init__.py
@@ -1,4 +1,4 @@
-"""Quantfolio is a Python package for portfolio optimization and risk management."""
+"""CoreFolio is a Python package for portfolio optimization and risk management."""
 
 from .universe import Universe
 from .constraints import Constraints
